tidy up helper fns, switch to using parsers
